+++ cmd/users/search_test.go
@@ -8,7 +8,7 @@ import (
 	"testing"
 
 	"github.com/flant/glaball/pkg/limiter"
-	"github.com/flant/glaball/pkg/sort"
+	"github.com/flant/glaball/pkg/sort/v2"
 
 	"github.com/flant/glaball/cmd/common"
