+++ cmd/versions/main.go
@@ -11,7 +11,7 @@ import (
 
 	"github.com/flant/glaball/pkg/client"
 	"github.com/flant/glaball/pkg/limiter"
-	"github.com/flant/glaball/pkg/sort"
+	"github.com/flant/glaball/pkg/sort/v2"
 
 	"github.com/flant/glaball/cmd/common"
